requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
deap==1.3.3
decorator==5.1.1
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
joblib==1.3.1
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
msgpack==1.0.5
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pooch==1.6.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyAudioAnalysis==0.3.14
pycparser==2.21
pyparsing==3.0.9
PySimpleGUI==4.33.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
stopit==1.1.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
TPOT==0.12.
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
update-checker==0.18.0
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
xgboost==1.7.6
zipp==3.15.0
